do optimize and cleanup

+++ src/main/java/teamport/aether/mixin/item/LavaBucketMixin.java
@@ -132,7 +132,7 @@ public abstract class LavaBucketMixin extends Item {
                 world.playSoundEffect(null, SoundCategory.WORLD_SOUNDS, (double) x + 0.5, (double) y + 0.5, (double) z + 0.5, "fire.ignite", 1.0F, itemRand.nextFloat() * 0.4F + 0.8F);
                 for (int l = 0; l < 8; ++l) {
                     double angle = Math.toRadians(l * 45);
-                    ParticleHelper.spawnParticle(world, "smoke", (double) x + 0.5, (double) y, (double) z + 0.5, -(Math.cos(angle) * 2) / 20.0, 0.03, -(Math.sin(angle) * 2) / 20.0, 0);
+                    ParticleHelper.spawnParticle(world, "smoke", (double) x + 0.5, y, (double) z + 0.5, -(Math.cos(angle) * 2) / 20.0, 0.03, -(Math.sin(angle) * 2) / 20.0, 0);
                 }
 
                 world.setBlockWithNotify(x, y, z, AetherBlocks.AEROGEL.id());
